Refactor source code

+++ examples/Speed-Calibrators/Wheel_Motor_Speed_Calculator-2/Motor_Rotation_Counter.h
@@ -15,7 +15,7 @@ class MotorRotationCounter {
 
   void handleInterrupt ();
 
-  volatile int counter_;
+  volatile uint16_t counter_;
 
   public:
     MotorRotationCounter (const byte which, const byte diskslots = 20);
